\docType{data}
\name{alert_eval}
\alias{alert_eval}
\title{Saved output from robustALERT()}
\format{Data frame with the output from robustALERT used in the vignette.}
\description{
Summarized performance of multiple ALERT rules applied to
the fluData dataset. Desiged for quick loading for
vignette.
}
\examples{
data(alert_eval)
}
\keyword{alert_eval}

