% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{fluData}
\alias{fluData}
\title{ALERT test data}
\format{A weekly time series.}
\description{
Weekly influenza cases from a hospital
}
\examples{
data(fluData)

## project the proper start date for the upcoming flu season
createALERT(data=fluData, firstMonth=8, lag=7, minWeeks=8, allThresholds=TRUE, k=2, target.pct=0.85)

## cross validate the above projection
robustALERT(minPercent=c(.8, .85, .9), maxDuration=c(12, 13, 14), data=fluData, firstMonth=8, lag=7, minWeeks=8, allThresholds=TRUE, k=2, target.pct=0.85)
}
\keyword{fluData}

